Added ip validator and added user register page.

+++ src/main/resources/templates/admin/user_list.ftl
@@ -33,7 +33,7 @@
 			<td>${user.enabled?then("Y", "N")}</td>
 			<td><#list user.ips as ip>${ip.ip}<#sep>, </#sep></#list></td>
 			<td><#list user.authorities as authority>${authority.authority}<#sep>, </#sep></#list></td>
-			<th><a href="/admin/user/delete/${user.username}">Delete</a></td>
+			<th><a onclick="javascript:if (!confirm('Delete?')) return false;" href="/admin/user/delete/${user.username}">Delete</a></td>
 		</tr>
 		</#list>
 	</table>
@@ -49,7 +49,7 @@
 	</form>
 
 	<div>
-		<a href="/admin/user/form/">Add user</a>
+		<a href="/admin/user/form/new">Add user</a>
 	</div>
 </body>
 </html>
